Polish in Level, Created Main Menu, Started Level Select
Tonight I polished the level so that it works better and looks better. It now has win, lose and mission UI. The game is able to tell if you beat the level. I have set up the functionality for having one scene and loading a particular level based of which level you select. I am using scriptable objects to hold the info that the game needs to set up the level. Tomorrow I am going to work on creating the level select menu.